Pre-release bugfixes for stock firmware probing
* Fix logic bug in WCS offset calculation.

Fusion allows us to shift the WCS relative to the probing point by proving an expected clearance value from the WCS. It turns out this clearance value is already compensating for the tool diameter. We were incorrectly also applying this compensation which is now fixed.

* Adds logic for first probe operation:

During an initial probe-z, Fusion will move to a clearance height relative to the current WCS coordinates in the machine. This was an unprotected rapid move, which could expose the probe to a crash if a workpiece of sufficiently different size were loaded.

To mitigate this we introduce the following logic:
- if this is the first operation, and it is a probe-z operation
- onSection: do not move to the operation Z height
- onRapid: do not move to the operation Z height
- onCyclePoint: do not move to the operation Z height
- onCyclePoint probe-z handler: probe Z-120

Since new programs can be assured to start at the G28 clerance Z height, this means the probe will be able to make a G38.3/5 safe move to the clearance height, stopping if it detects a collision.

This code is complicated by the fact tat G38.x moves are purely relative. Therefore we can't simply change the rapid moves to a G38.3/5 because the post-processor is unaware of the starting coordinates.

* Introduce handler.absoluteRapidMove to speed up retractions at the end of probing cycles. This was handled by onSection before but not as a rapid.

* Set double-tap retraction to tool.diameter instead of 2x diameter, since the extra distance was not necessary

* Tweak gFormat to support decimal places for G38 etc. Remove gCycleModal to keep changes at a minimum

* Rename handler.relativeMove to relativeSafeMove for clarity

* Add documentation references for Fusion's probing docs