implement optional 4th Axis rotation based on Angle between WCS and Model
This allows you to define angled WCS in setups and then automatically adds the gcode for A axis rotation. Make sure the Model Coordinate System is aligned with the carvera Global Coordinate System. This means your part's x axis has to be the rotation Axis. Rotate your Part if necessary.